refactor: move churn trend contract to types crate
